+++ sdk/kusto/arm-kusto/src/models/index.ts
@@ -90,6 +90,7 @@ export type {
   OptimizedAutoscale,
   VirtualNetworkConfiguration,
   VnetState,
+  LanguageExtensionsList,
   LanguageExtension,
   LanguageExtensionName,
   LanguageExtensionImageName,